+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryClientModule.java
@@ -132,8 +132,6 @@ public class BigQueryClientModule implements com.google.inject.Module {
             .build();
     return new BigQueryClient(
         options.build().getService(),
-        config.getMaterializationProject(),
-        config.getMaterializationDataset(),
         destinationTableCache,
         bigQueryJobLabels,
         config.getQueryJobPriority(),