+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/ReadSessionCreatorConfig.java
@@ -24,10 +24,7 @@ import java.util.OptionalLong;
 
 public class ReadSessionCreatorConfig {
   private final boolean viewsEnabled;
-  private final Optional<String> materializationProject;
-  private final Optional<String> materializationDataset;
   private final String viewEnabledParamName;
-  private final int materializationExpirationTimeInMinutes;
   private final DataFormat readDataFormat;
   private final int maxReadRowsRetries;
   private final OptionalInt maxParallelism;
@@ -49,9 +46,6 @@ public class ReadSessionCreatorConfig {
 
   ReadSessionCreatorConfig(
       boolean viewsEnabled,
-      Optional<String> materializationProject,
-      Optional<String> materializationDataset,
-      int materializationExpirationTimeInMinutes,
       DataFormat readDataFormat,
       int maxReadRowsRetries,
       String viewEnabledParamName,
@@ -72,10 +66,7 @@ public class ReadSessionCreatorConfig {
       long readSessionCacheDurationMins,
       OptionalLong snapshotTimeMillis) {
     this.viewsEnabled = viewsEnabled;
-    this.materializationProject = materializationProject;
-    this.materializationDataset = materializationDataset;
     this.viewEnabledParamName = viewEnabledParamName;
-    this.materializationExpirationTimeInMinutes = materializationExpirationTimeInMinutes;
     this.readDataFormat = readDataFormat;
     this.maxReadRowsRetries = maxReadRowsRetries;
     this.maxParallelism = maxParallelism;
@@ -100,20 +91,8 @@ public class ReadSessionCreatorConfig {
     return viewsEnabled;
   }
 
-  public Optional<String> getMaterializationProject() {
-    return materializationProject;
-  }
-
-  public Optional<String> getMaterializationDataset() {
-    return materializationDataset;
-  }
-
   public String getViewEnabledParamName() {
     return viewEnabledParamName;
-  }
-
-  public int getMaterializationExpirationTimeInMinutes() {
-    return materializationExpirationTimeInMinutes;
   }
 
   public DataFormat getReadDataFormat() {