+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SparkBigQueryConfig.java
@@ -47,7 +47,6 @@ import com.google.cloud.bigquery.connector.common.BigQueryConnectorException;
 import com.google.cloud.bigquery.connector.common.BigQueryCredentialsSupplier;
 import com.google.cloud.bigquery.connector.common.BigQueryProxyConfig;
 import com.google.cloud.bigquery.connector.common.BigQueryUtil;
-import com.google.cloud.bigquery.connector.common.MaterializationConfiguration;
 import com.google.cloud.bigquery.connector.common.ReadSessionCreatorConfig;
 import com.google.cloud.bigquery.connector.common.ReadSessionCreatorConfigBuilder;
 import com.google.cloud.bigquery.storage.v1.ArrowSerializationOptions.CompressionCodec;
@@ -202,8 +201,6 @@ public class SparkBigQueryConfig
   DataFormat readDataFormat = DEFAULT_READ_DATA_FORMAT;
   boolean combinePushedDownFilters = true;
   boolean viewsEnabled = false;
-  com.google.common.base.Optional<String> materializationProject = empty();
-  com.google.common.base.Optional<String> materializationDataset = empty();
   com.google.common.base.Optional<String> partitionField = empty();
   Long partitionExpirationMs = null;
   com.google.common.base.Optional<Boolean> partitionRequireFilter = empty();
@@ -217,7 +214,6 @@ public class SparkBigQueryConfig
   boolean useAvroLogicalTypes = false;
   List<String> decimalTargetTypes = Collections.emptyList();
   List<JobInfo.SchemaUpdateOption> loadSchemaUpdateOptions = Collections.emptyList();
-  int materializationExpirationTimeInMinutes;
   int maxReadRowsRetries = 3;
   boolean pushAllFilters = true;
   boolean enableModeCheckForSchemaFields = true;
@@ -343,22 +339,12 @@ public class SparkBigQueryConfig
     ImmutableMap<String, String> globalOptions = normalizeConf(originalGlobalOptions);
     config.sparkBigQueryProxyAndHttpConfig =
         SparkBigQueryProxyAndHttpConfig.from(options, globalOptions, hadoopConfiguration);
-    // Issue #247
-    // we need those parameters in case a read from query is issued
-    config.viewsEnabled = getAnyBooleanOption(globalOptions, options, VIEWS_ENABLED_OPTION, false);
-    MaterializationConfiguration materializationConfiguration =
-        MaterializationConfiguration.from(globalOptions, options);
-    config.materializationProject = materializationConfiguration.getMaterializationProject();
-    config.materializationDataset = materializationConfiguration.getMaterializationDataset();
-    config.materializationExpirationTimeInMinutes =
-        materializationConfiguration.getMaterializationExpirationTimeInMinutes();
     // get the table details
-    com.google.common.base.Optional<String> fallbackDataset = config.materializationDataset;
     Optional<String> fallbackProject =
         com.google.common.base.Optional.fromNullable(
                 hadoopConfiguration.get(GCS_CONFIG_PROJECT_ID_PROPERTY))
             .toJavaUtil();
-    Optional<String> datasetParam = getOption(options, "dataset").or(fallbackDataset).toJavaUtil();
+    Optional<String> datasetParam = getOption(options, "dataset").toJavaUtil();
     Optional<String> projectParam =
         firstPresent(getOption(options, "project").toJavaUtil(), fallbackProject);
     config.partitionType =
@@ -937,16 +923,6 @@ public class SparkBigQueryConfig
     return viewsEnabled;
   }
 
-  @Override
-  public Optional<String> getMaterializationProject() {
-    return materializationProject.toJavaUtil();
-  }
-
-  @Override
-  public Optional<String> getMaterializationDataset() {
-    return materializationDataset.toJavaUtil();
-  }
-
   public Optional<String> getPartitionField() {
     return partitionField.toJavaUtil();
   }
@@ -997,10 +973,6 @@ public class SparkBigQueryConfig
 
   public ImmutableList<JobInfo.SchemaUpdateOption> getLoadSchemaUpdateOptions() {
     return ImmutableList.copyOf(loadSchemaUpdateOptions);
-  }
-
-  public int getMaterializationExpirationTimeInMinutes() {
-    return materializationExpirationTimeInMinutes;
   }
 
   public int getMaxReadRowsRetries() {
@@ -1154,9 +1126,6 @@ public class SparkBigQueryConfig
   public ReadSessionCreatorConfig toReadSessionCreatorConfig() {
     return new ReadSessionCreatorConfigBuilder()
         .setViewsEnabled(viewsEnabled)
-        .setMaterializationProject(materializationProject.toJavaUtil())
-        .setMaterializationDataset(materializationDataset.toJavaUtil())
-        .setMaterializationExpirationTimeInMinutes(materializationExpirationTimeInMinutes)
         .setReadDataFormat(readDataFormat)
         .setMaxReadRowsRetries(maxReadRowsRetries)
         .setViewEnabledParamName(VIEWS_ENABLED_OPTION)
@@ -1199,11 +1168,6 @@ public class SparkBigQueryConfig
       @Override
       public String viewEnabledParamName() {
         return SparkBigQueryConfig.VIEWS_ENABLED_OPTION;
-      }
-
-      @Override
-      public int expirationTimeInMinutes() {
-        return SparkBigQueryConfig.this.getMaterializationExpirationTimeInMinutes();
       }
     };
   }