+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SparkBigQueryUtil.java
@@ -198,7 +198,10 @@ public class SparkBigQueryUtil {
   public static TableId parseSimpleTableId(SparkSession spark, Map<String, String> options) {
     ImmutableMap<String, String> globalOptions =
         ImmutableMap.copyOf(scalaMapToJavaMap(spark.conf().getAll()));
-    return BigQueryConfigurationUtil.parseSimpleTableId(globalOptions, options);
+    return BigQueryConfigurationUtil.parseSimpleTableId(
+        options,
+        com.google.common.base.Optional.absent(),
+        com.google.common.base.Optional.absent());
   }
 
   public static long sparkTimestampToBigQuery(Object sparkValue) {