+++ spark-bigquery-connector-common/src/test/java/com/google/cloud/spark/bigquery/integration/ReadFromQueryIntegrationTestBase.java
@@ -72,13 +72,7 @@ class ReadFromQueryIntegrationTestBase extends SparkBigQueryIntegrationTestBase 
   }
 
   private void testReadFromQueryInternal(String query) {
-    Dataset<Row> df =
-        spark
-            .read()
-            .format("bigquery")
-            .option("viewsEnabled", true)
-            .option("materializationDataset", testDataset.toString())
-            .load(query);
+    Dataset<Row> df = spark.read().format("bigquery").option("viewsEnabled", true).load(query);
 
     validateResult(df);
     // validate event publishing
@@ -123,13 +117,7 @@ class ReadFromQueryIntegrationTestBase extends SparkBigQueryIntegrationTestBase 
             "SELECT corpus, word_count FROM `bigquery-public-data.samples.shakespeare` WHERE word='spark' AND '%s'='%s'",
             random, random);
     Dataset<Row> df =
-        spark
-            .read()
-            .format("bigquery")
-            .option("viewsEnabled", true)
-            .option("materializationDataset", testDataset.toString())
-            .option("query", query)
-            .load();
+        spark.read().format("bigquery").option("viewsEnabled", true).option("query", query).load();
 
     StructType expectedSchema =
         DataTypes.createStructType(
@@ -181,12 +169,7 @@ class ReadFromQueryIntegrationTestBase extends SparkBigQueryIntegrationTestBase 
     assertThrows(
         RuntimeException.class,
         () -> {
-          spark
-              .read()
-              .format("bigquery")
-              .option("viewsEnabled", true)
-              .option("materializationDataset", testDataset.toString())
-              .load(badSql);
+          spark.read().format("bigquery").option("viewsEnabled", true).load(badSql);
         });
   }
 
@@ -202,7 +185,6 @@ class ReadFromQueryIntegrationTestBase extends SparkBigQueryIntegrationTestBase 
             .read()
             .format("bigquery")
             .option("viewsEnabled", true)
-            .option("materializationDataset", testDataset.toString())
             .option("queryJobPriority", "batch")
             .load(query);
 
@@ -220,7 +202,6 @@ class ReadFromQueryIntegrationTestBase extends SparkBigQueryIntegrationTestBase 
                 .read()
                 .format("bigquery")
                 .option("viewsEnabled", true)
-                .option("materializationDataset", testDataset.toString())
                 .option("bigQueryJobTimeoutInMinutes", "1")
                 .load(query)
                 .show();