+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 * PR #1284: Making BigQueryClientFactory Kryo serializable. Thanks @tom-s-powell !
+* PR #XXXX: `materializationDataset` is no longer needed to read from views or queries.
 
 ## 0.42.0 - 2025-02-06
 * PR #1333: Initial implementation of a BigQuery backed Spark Catalog